+++ netket/nn/__init__.py
@@ -37,7 +37,6 @@ from .activation import (
     log_tanh,
     reim_selu,
     reim_relu,
-    logsumexp_cplx,
 )
 from flax.linen import (
     MultiHeadDotProductAttention,
